% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appreciate.R
\name{appreciate}
\alias{appreciate}
\title{Appreciate your package dependencies}
\usage{
appreciate()
}
\value{
Always returns an invisible data.frame with three columns "username", "repository" and "starred".
Also outputs that information to the console.
}
\description{
This function is meant to be used when developing a package.
It extracts all package dependencies from the \code{DESCRIPTION} file in your working directory
and checks if you have starred it on github (if those packages are hosted on github).
}
\details{
Still under development. Api and names can change.
}
\references{
Inspired by <https://github.com/musically-ut/appreciate>
}
