% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_starred.R
\name{gh_starred}
\alias{gh_starred}
\title{Check if you starred packages on Github}
\usage{
gh_starred(packages, console_output = TRUE,
  api_url = "https://api.github.com",
  user_agent = "https://github.com/dirkschumacher/thankr")
}
\arguments{
\item{packages}{a non empty character vector of package names.}

\item{console_output}{Optional. TRUE if the result should be written to the console}

\item{api_url}{Optional. The GitHub API url.}

\item{user_agent}{Optional. The user-agent for all the API requests.
Change it if this function is part of something bigger.}
}
\value{
Always returns a data.frame with three columns "username", "repository" and "starred".
If \code{console_output} is TRUE, then the return value is invisible.
}
\description{
Starring a package on Github is a form of appreciation for the project.
With this function you can quickly check if you starred all your favorite packages.
}
\details{
The function assumes that you have set an environment variable
(either \code{GITHUB_PAT} or \code{GITHUB_TOKEN}) with a Github Access Token.

When using this function together with \code{\link{shoulders}} please note that
once run for the first time additional packages are attached to your session.

Also take a look at the API documentation <https://developer.github.com/v3/> of Github
and their terms of use before using this function.
}
\references{
Inspired by <https://github.com/musically-ut/appreciate>
}
