% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoulders.R
\name{shoulders}
\alias{shoulders}
\title{Find out who wrote the packages you use}
\usage{
shoulders(where = c("namespace", "library", "package"), package = NULL,
  include_dependencies = TRUE)
}
\arguments{
\item{where}{either look in the current loaded "namespace", your whole "library" or for a specific package}

\item{package}{if \code{where} is "package" then this must be a character vector of positive length}

\item{include_dependencies}{if \code{where} is "package" then use this to include all dependencies of the package.
Otherwise it will just return the maintainer.}
}
\value{
a tibble with three columns, \code{maintainer}, \code{no_packages}, \code{packages} showing all packages you depend on.
}
\description{
This function helps to identify what packages you depend on. It identifies the maintainers and the
number of packages they wrote.
}
\details{
If you do a package lookup you connect to a CRAN mirror.
}
