% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appreciate.R
\name{appreciate}
\alias{appreciate}
\title{Appreciate your package dependencies}
\usage{
appreciate()
}
\value{
Always returns an invisible data.frame with three columns "username", "repository" and "starred".
Also outputs that information to the console.
}
\description{
This function is meant to be used when developing a package.
It extracts all package dependencies from the \code{DESCRIPTION} file in your working directory
and checks if you have starred it on github (if those packages are hosted on github).
}
\details{
Still under development. Api and names can change.

The function assumes that you have set an environment variable
(either \code{GITHUB_PAT} or \code{GITHUB_TOKEN}) with a Github Access Token.

When using this function together with \code{\link{shoulders}} please note that
once run for the first time additional packages are attached to your session.

Also take a look at the API documentation <https://developer.github.com/v3/> of Github
and their terms of use before using this function.
}
\references{
Inspired by <https://github.com/musically-ut/appreciate>
}
