% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thankr-package.R
\docType{package}
\name{thankr}
\alias{thankr}
\alias{thankr-package}
\title{Find Out Who Maintains the Packages you Use}
\description{
Find out who maintains the packages you use in
your current session or in your package library.
}
\details{
This package exposes a single function \code{\link[thankr]{shoulders}} that helps
you identify the maintainers of your favorite packages.

> standing on the shoulders of giants
}
\keyword{package}
